Refactor: "자유게시판 전체조회 리팩토링"
자유게시판 전체조회 시 불필요한 정보는 표시되지 않게 변경

Resolves: #33
