+++ itty-backend/src/main/java/org/iot/itty/article/vo/ResponseReplyOfArticleList.java
@@ -1,0 +1,8 @@
+package org.iot.itty.article.vo;
+
+import lombok.Data;
+
+@Data
+public class ResponseReplyOfArticleList {
+	private int replyCodePk;
+}
