+++ itty-backend/src/main/java/org/iot/itty/user/vo/ResponseAuthorOfArticleList.java
@@ -1,0 +1,9 @@
+package org.iot.itty.user.vo;
+
+import lombok.Data;
+
+@Data
+public class ResponseAuthorOfArticleList {
+	private int userCodePk;
+	private String userNickname;
+}
